Move launcher CLI to main CLI
