+++ tests/tests_app/launcher/test_running_flow.py
@@ -9,7 +9,7 @@ import pytest
 import requests
 
 from lightning.app.launcher import launcher, lightning_backend
-from lightning.app.launcher.utils import convert_print_to_logger_info
+from lightning.app.utilities.app_helpers import convert_print_to_logger_info
 from lightning.app.utilities.enum import AppStage
 from lightning.app.utilities.exceptions import ExitAppException
